WHITE PAPER 7/11/24
This is the official White Paper of Deez Nuts.
A Meme token on the Algorand Blockchain launching on https://rug.ninja/Deez

Join the undead squirrel's journey for Deez Nutz, glazed with BTC rewards on the Algorand blockchain.
As of 2024, there are over 30 million unique addresses (wallets) on the Algorand blockchain. Ergo, max supply: 2 Nutz times the (Algo wallets) = 69 million in total.